generative_model_training/requirements.txt
einops==0.6.1
facenet_pytorch==2.5.3
hydra-core==1.3.2
matplotlib==3.7.2
mxnet==1.9.1
numpy==1.24.4
omegaconf==2.3.0
pandas==2.0.3
Pillow==10.0.0
pyeer==0.5.5
pytorch_lightning==1.5.*
Requests==2.31.0
scikit_learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
scikit-image==0.21.0
tqdm==4.66.1
